requirements.txt
flask==2.2.2
matplotlib==3.6.2
itsdangerous==2.1.2
mysql-connector-python==8.0.31
pandas==1.5.2
basemap==1.3.6
requests==2.28.1